+++ Modules/Microsoft365DSC/DscResources/MSFT_AADConditionalAccessPolicy/MSFT_AADConditionalAccessPolicy.psm1
@@ -233,7 +233,7 @@ class AADConditionalAccessPolicy : M365DSCResourceBase
 
     [DscProperty()]
     [System.ComponentModel.Description('Service principal risk levels included in the policy.')]
-    [ValidateSet('low', 'medium', 'high', 'none', 'unknownFutureValue')]
+    [ValidateSet('low', 'medium', 'high', 'none', 'unknownFutureValue', 'hidden')]
     [System.String[]] $ServicePrincipalRiskLevels
 
     [DscProperty()]